parser/requirements.txt
requests
bs4
upstash-redis
python-dotenv